ia-audio/requirements.txt
flask
openai-whisper
torch
numpy
